Fix serialization problems
In previous fix launcher was introduced to the abstract methods in
RemoteAgent. Due to proxy objects, for some of the sub-classes of that
class this results in unforeseen serialization problems caused by one
of the arguments which is unserializable (Launcher).

This change removes that argument from the affected methods, instead it
is propagated using a provider pattern for those that need it.