Support execution on Windows by using ssh-agent provided by git
Since Windows-10 a ssh-agent is installed by default on Windows systems.
But that agent works differently than the ssh-agent known from UNIX
systems and therefore attempts to use this plugin fail.
From the error message the cause of the failure is not easy to
understand. And of course not being able to use this plugin on Windows
agents is in general a unsatisfactory situation.

Fortunately the git (for-windows) installation by default contains a
ssh-agent implementation that behaves like the known UNIX implementation
and runs on Windows. And a git installation is often available.

This change extends the ssh-agent plugin to attempt to locate the
described ssh-agent installation provided by git, when running on
Windows.
If no executable is found, either through a git installation or by
searching the system's PATH variable, the execution fails with a
meaningful message.